pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vendor-performance-analytics"
version = "2.0.0"
description = "End-to-end retail analytics pipeline for vendor performance evaluation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Aman Saroha", email = "amaninsights@github.com"}
]
keywords = [
    "data-analytics",
    "vendor-analysis",
    "retail-analytics",
    "business-intelligence",
    "data-pipeline",
    "sql",
    "etl"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
requires-python = ">=3.9"
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "scipy>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.scripts]
vendor-analytics = "main:main"

[project.urls]
Homepage = "https://github.com/amaninsights/Vendor-Performance-Analysis"
Repository = "https://github.com/amaninsights/Vendor-Performance-Analysis"
Documentation = "https://github.com/amaninsights/Vendor-Performance-Analysis#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
